requirements.txt
streamlit
yfinance
plotly
pandas
requests
python-dotenv
alpaca-py
feedparser
